refactor(app): extract HomeComposer out of HomePage
HomePage.tsx was 1,183 lines and sitting at its documented modularity
budget (1,190). The inline HomeComposer — the launchpad prompt box with
its model/reasoning/assign menus, attachments and mention picker — was
the bulk of it and the named item in the file's extraction backlog.

Move HomeComposer into components/home/HomeComposer.tsx (553 lines) and
lift the shared primary-agent-mode helpers (allowedPrimaryAgentModes,
constrainedPrimaryAgentMode, nextAllowedPrimaryAgentMode) into the
existing helpers/primary-agent-mode.ts — the canonical module for that
domain — so HomeComposer never imports from HomePage (no cycle) and we
avoid a second same-named file.

HomePage drops to 621 lines; its budget is ratcheted 1,190 -> 630.
Behaviour-preserving: renderer tests, typecheck, eslint, knip, and the
import-cycle guard all pass.
